Way better timer
Scripted a more efficient timer that works with TMP.

This timer doesn't use the update function like my first one did, instead it uses a coroutine function to be called at start of script instead.

In my final proj this will be triggered by raycast hit detection on the first blow to the enemy ship.